ARM: dts: am335x-evmsk: fix wilink from blocking suspend
wlcore driver requires the SDIO controller to support
keeping the card powered during suspend.

According to a comment in the driver,
this is required to workaround issues with power sequencing
and wlcore driver's preference is to control card power itself.

Without this patch,
AM335x EVM-SK refuses to suspend while wlcore driver is installed.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Yaniv Machani <yanivma@ti.com>